move extrapolateTrackPositionToSensor to TrackUtils
